camera/pyproject.toml
[tool.poetry]
name = "shuffleboard"
version = "0.1.0"
description = ""
authors = ["Ned Reid"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
opencv-python = "^4.10.0.84"
python-socketio = "^5.11.4"
asyncio = "^3.4.3"
aiohttp = "^3.11.2"
dataclasses-json = "^0.6.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
